requirements.txt
langchain
GitPython
